backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
redis==5.2.1
Pillow==11.1.0
python-dotenv==1.0.1
PyJWT==2.10.1
whitenoise==6.8.2
